[ECO-5380] Create separated unit test file for ObjectMessage 1. Added three different unit tests covering edge cases 2. Updated doc for byteSize String extension method 3. Updated type of maxMessageSize to int instead of long

+++ lib/src/main/java/io/ably/lib/objects/Adapter.java
@@ -31,7 +31,7 @@ public class Adapter implements LiveObjectsAdapter {
     }
 
     @Override
-    public long maxMessageSizeLimit() {
+    public int maxMessageSizeLimit() {
         return ably.connection.connectionManager.maxMessageSize;
     }
 }
